Extract orchestrator support modules
